requirements.txt
# ====================================
# Core Dependencies
# ====================================
python-dotenv==1.1.0
requests>=2.31.0
numpy==1.26.4

# ====================================
# Web Interface
# ====================================
gradio==4.44.1
gradio-client==1.3.0
gradio-webrtc==0.0.9a1
ipywidgets==8.1.6

# ====================================
# PyTorch (CPU version for compatibility)
# ====================================
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0

# ====================================
# Hugging Face Transformers for Whisper STT
# ====================================
transformers>=4.35.0
tokenizers==0.21.1
huggingface-hub>=0.30.0
datasets==3.5.0
accelerate

# ====================================
# Audio Processing
# ====================================
librosa==0.11.0
soundfile==0.13.1
scipy==1.15.2

# ====================================
# Text-to-Speech (Kokoro)
# ====================================
phonemizer==3.3.0
munch==4.0.0
misaki>=0.7.0

# ====================================
# AWS and NVIDIA NIM
# ====================================
boto3>=1.28.0
sagemaker>=2.200.0

# ====================================
# RAG and Embeddings
# ====================================
scikit-learn>=1.3.0

# ====================================
# Utilities
# ====================================
spaces==0.35.0
xxhash==3.5.0